# maths: switch dot to faster for loop