Tiny optimization; the for loop warms up faster than the reduce.

```js
import { dot } from "@xenova/transformers";

function dotNew(arr1, arr2) {
    let result = 0;
    for (let i = 0; i < arr1.length; i++) {
        result += arr1[i] * arr2[i];
    }
    return result;
}

const random1 = Array.from({ length: 10000 }, () => Array.from({ length: 512 }, () => Math.random()));
const random2 = Array.from({ length: 10000 }, () => Array.from({ length: 512 }, () => Math.random()));

// Replaced with dotNew for second run
console.time("dot");
for (let i = 0; i < 10000; i++) {
    dot(random1[i], random2[i]);
}
console.timeEnd("dot");
```

Output on `node v20.5.1`

```
dot: 43.757ms
dotNew: 11.982ms
```

Of course, eventually they're exactly the same. With 100k elements:
```
dot: 109.57ms
dotNew: 76.428ms
```