@xenova 2024-03-27
Oh interesting! Thanks for this! Could you do some benchmarks with typed arrays? Specifically, `Float32Array`?