@pulsejet 2024-03-27
Well the current dot function doesn't work at all on `Float32Array` (another reason to switch to the new one) because there's no `reduce`.

Typed arrays are extremely fast.

```
dot: 43.757ms
dotNew: 11.982ms
dotNew: 0.887ms  (Float32Array)
dotNew: 0.919ms (Float64Array)
```

With 100k elements
```
dot: 109.57ms
dotNew: 76.428ms
dotNew: 3.237ms (Float32Array)
dotNew: 3.167ms (Float64Array)
```