@xenova 2024-04-10
Just for the fun of it, I was doing some benchmarking myself (both in browser and node), and your update is consistently 5-10x faster (of course, depends on the array size)! Honestly, I'm quite surprised, since I would expect this to be a common pattern, and for `reduce` to be optimized for this case.

Thanks!